requirements.txt
plotly==4.5.0
numpy==1.19.5
pandas==1.2.2
Django==3.2.1
django-picklefield==3.0.1
scikit_learn==0.24.2
